Refactor import statements: update import paths for consistency and clarity across modules.
